move all components to src/components, and only document them
